requirements.txt
asgiref==3.4.1
bcrypt==3.2.0
cffi==1.14.6
click==8.0.1
colorama==0.4.4
cryptography==3.4.8
dnspython==1.16.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.68.1
h11==0.12.0
httptools==0.2.0
idna==3.2
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
pymongo==3.12.0
python-dotenv==0.19.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==5.4.1
rsa==4.7.2
six==1.16.0
starlette==0.14.2
typing-extensions==3.10.0.2
uvicorn==0.15.0
watchgod==0.7
websockets==10.0
